Get leaflet polygon from ui-components
