+++ packages/ui-components/src/components/Map/index.js
@@ -3,13 +3,14 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  *
  */
+export { Polygon } from 'react-leaflet';
 export * from './AreaTooltip';
+export * from './EntityPolygon';
 export * from './InversePolygonMask';
 export * from './Legend';
 export * from './MapContainer';
 export * from './MarkerLayer';
 export * from './Markers';
-export * from './Polygon';
 export * from './PolygonLayer';
 export * from './TileLayer';
 export * from './TilePicker';
